refactor(tasks): move Sharpa in-hand runtime [skip ci]
