+++ scripts/benchmark/core/task_names.py
@@ -2,10 +2,10 @@ from __future__ import annotations
 
 from dataclasses import dataclass
 
-from unilab.envs.manipulation.sharpa_inhand.rotation import SharpaInhandRotationCfg
 from unilab.tasks.locomotion.g1.joystick import G1WalkFlatCfg
 from unilab.tasks.locomotion.go1.joystick import Go1JoystickCfg
 from unilab.tasks.locomotion.go2.joystick import Go2JoystickCfg
+from unilab.tasks.manipulation.sharpa_inhand.rotation import SharpaInhandRotationCfg
 
 
 @dataclass(frozen=True)
