+++ tests/test_sharpa.py
@@ -5,12 +5,12 @@ from types import SimpleNamespace
 import numpy as np
 import pytest
 
-from unilab.envs.manipulation.sharpa_inhand.base import SharpaDomainRandConfig
-from unilab.envs.manipulation.sharpa_inhand.grasp_gen import (
+from unilab.tasks.manipulation.sharpa_inhand.base import SharpaDomainRandConfig
+from unilab.tasks.manipulation.sharpa_inhand.grasp_gen import (
     SharpaInhandRotationGraspCfg,
     SharpaInhandRotationGraspEnv,
 )
-from unilab.envs.manipulation.sharpa_inhand.rotation import (
+from unilab.tasks.manipulation.sharpa_inhand.rotation import (
     SharpaInhandRotationEnv,
     sample_random_quaternion,
 )
